Improve version/tag handling & display

+++ weblab/entities/views.py
@@ -24,6 +24,7 @@ from django.views.generic.list import ListView
 from git import GitCommandError
 
 from .forms import (
+    EntityTagVersionForm,
     EntityVersionForm,
     FileUploadForm,
     ModelEntityForm,
@@ -65,6 +66,7 @@ class VersionMixin:
     Mixin for including in pages describing a specific version
     of an `Entity` object
     """
+
     def get_context_data(self, **kwargs):
         entity = self.get_object()
         tags = entity.tag_dict
@@ -75,7 +77,7 @@ class VersionMixin:
             commit = entity.repo.commit(version)
         kwargs.update(**{
             'version': commit,
-            'tag': tags.get(commit),
+            'tags': tags.get(commit, []),
         })
         return super().get_context_data(**kwargs)
 
@@ -141,6 +143,7 @@ class EntityVisibilityMixin(AccessMixin, SingleObjectMixin):
     If an entity is not visible to a logged in user, we generate a 404
     If an entity is not visible to an anonymous visitor, redirect to login page
     """
+
     def dispatch(self, request, *args, **kwargs):
         # We don't necessarily want 'object not found' to give a 404 response
         # (if the user is anonymous it makes more sense to login-redirect them)
@@ -195,6 +198,48 @@ class EntityView(EntityVisibilityMixin, SingleObjectMixin, RedirectView):
     def get_redirect_url(self, *args, **kwargs):
         url_name = 'entities:{}_version'.format(kwargs['entity_type'])
         return reverse(url_name, args=[kwargs['pk'], 'latest'])
+
+
+class EntityTagVersionView(
+    LoginRequiredMixin, FormMixin, VersionMixin, DetailView
+):
+    """Add a new tag to an existing version of an entity."""
+    context_object_name = 'entity'
+    form_class = EntityTagVersionForm
+    model = Entity
+    template_name = 'entities/entity_tag_version.html'
+
+    def get_context_data(self, **kwargs):
+        entity = self.get_object()
+        kwargs['type'] = entity.entity_type
+        return super().get_context_data(**kwargs)
+
+    def post(self, request, *args, **kwargs):
+        """Check the form and possibly add the tag in the repo.
+
+        Called by Django when a form is submitted.
+        """
+        import git
+        form = self.get_form()
+        entity = self.object = self.get_object()
+        if form.is_valid():
+            version = self.kwargs['sha']
+            tag = form.cleaned_data['tag']
+            try:
+                entity.tag_repo(tag, ref=version)
+            except git.exc.GitCommandError as e:
+                msg = e.stderr.strip().split(':', 1)[1][2:-1]
+                form.add_error('tag', msg)
+                return self.form_invalid(form)
+            return self.form_valid(form)
+        else:
+            return self.form_invalid(form)
+
+    def get_success_url(self):
+        """What page to show when the form was processed OK."""
+        entity = self.get_object()
+        version = self.kwargs['sha']
+        return reverse('entities:%s_version' % entity.entity_type, args=[entity.id, version])
 
 
 class EntityDeleteView(UserPassesTestMixin, DeleteView):
@@ -282,13 +327,14 @@ class EntityNewVersionView(
                 request.user.full_name,
                 request.user.email
             )
-            try:
-                entity.tag_repo(request.POST['version'])
-            except GitCommandError as e:
-                entity.repo.head.reset('HEAD~')
-                for f in entity.repo.untracked_files:
-                    os.remove(str(entity.repo_abs_path / f))
-                return self.fail_with_git_errors([e.stderr])
+            if request.POST['tag']:
+                try:
+                    entity.tag_repo(request.POST['tag'])
+                except GitCommandError as e:
+                    entity.repo.head.reset('HEAD~')
+                    for f in entity.repo.untracked_files:
+                        os.remove(str(entity.repo_abs_path / f))
+                    return self.fail_with_git_errors([e.stderr])
 
             # Temporary upload files have been safely committed, so can be deleted
             for filename in files_to_delete:
